requirements.txt

Django==5.1.2
django-admin-rangefilter==0.13.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-filter==24.2
django-import-export==4.3.3
django-leaflet==0.31.0
django-phonenumber-field==8.0.0
django-timezone-field==7.0
django-widget-tweaks==1.5.0
whitenoise==6.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-filter==24.2
django-htmx==1.21.0
django-import-export==4.3.3
phonenumberslite==8.13.51
gunicorn==22.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
dj-database-url==2.3.0
python-decouple==3.8
python-dotenv==1.0.1